server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.62.0",
    "matplotlib>=3.10.5",
    "mcp>=1.12.4",
    "numpy>=2.3.2",
    "python-dotenv>=1.1.1",
    "torch>=2.8.0",
]
